Close story when escape key is pressed
When the story is shown in a modal popup element using the Simple Modal
plugin, we want the modal to be closed when the escape key is pressed, to
offer similar behavior to when the story is shown in fullscreen.

We have to listen for the keystroke event in the story iframe window because
the parent loses focus and doesn't get the event.  Then we have to use
window.postMessage to inform the parent window. I'm not crazy about this,
but it seems to work.